Fix repo name in json files
The repo name was changed a few commits ago, forgot to update these json
files.  I thought this might've been what is causing a failure when
trying to pipe jest output to coveralls, but this still doesn't fix the
coveralls issue, so I'm not sure.

Here's the coveralls error I'm seeing:

/home/travis/build/robert-git/basketball_stats/node_modules/coveralls/bin/coveralls.js:18
        throw err;
                ^
Bad response: 422 {"message":"Couldn't find a repository matching this job.","error":true}

Looking at the error, it might be that "basketball_stats" is now
"basketball-stats".  Where does that need to be updated?