Implement eFG% stub, fix test, fix failing CI
The tests pass, but Travis-CI is still failing due to the coveralls
failure.

EDIT: I finally got coveralls to work with Travis-CI.  It seems to have
been an issue with the repo name.  A few commits ago I renamed the repo
from basketball_stats to basketball-stats (underscore to hypen), but
Travis-CI was still referencing the old name.  So I removed the repo
from Travis, then added it back, and everything is now passing.